+++ apps/ledger-live-desktop/src/renderer/components/EarnRewardsInfoModal.tsx
@@ -4,7 +4,6 @@ import styled from "styled-components";
 import Check from "../icons/CheckFull";
 import TrackPage from "../analytics/TrackPage";
 import Rewards from "../images/rewards.svg";
-import useTheme from "../hooks/useTheme";
 import Text from "./Text";
 import Button from "./Button";
 import Box from "./Box";
@@ -27,7 +26,6 @@ export default function EarnRewardsInfoModal({
   additional,
   footerLeft,
 }: Props) {
-  const infoColor = useTheme("colors.positiveGreen");
   return (
     <Modal
       name={name}
@@ -59,7 +57,7 @@ export default function EarnRewardsInfoModal({
                 <Box>
                   {bullets.map((val, i) => (
                     <Row key={val + i}>
-                      <Check size={16} color={infoColor} />
+                      <Check size={16} />
                       <Text
                         ff="Inter|SemiBold"
                         style={{