+++ apps/ledger-live-desktop/src/renderer/components/Onboarding/Screens/SelectUseCase/index.tsx
@@ -2,6 +2,8 @@ import React, { useContext } from "react";
 import { useHistory } from "react-router-dom";
 import { useTranslation, Trans } from "react-i18next";
 import { Flex, Text } from "@ledgerhq/react-ui";
+import { DeviceModelId } from "@ledgerhq/devices";
+import { useDispatch } from "react-redux";
 import styled from "styled-components";
 import { UseCaseOption } from "./UseCaseOption";
 import { ScrollArea } from "~/renderer/components/Onboarding/ScrollArea";
@@ -21,6 +23,7 @@ import restoreUsingRecoverDark from "./assets/restoreUsingRecoverDark.png";
 
 import Illustration from "~/renderer/components/Illustration";
 import { FeatureToggle } from "@ledgerhq/live-common/featureFlags/index";
+import { openModal } from "~/renderer/actions/modals";
 
 registerAssets([
   connectNanoLight,
@@ -84,6 +87,7 @@ export function SelectUseCase({ setUseCase, setOpenedPedagogyModal }: Props) {
   const { t } = useTranslation();
   const { deviceModelId } = useContext(OnboardingContext);
   const history = useHistory();
+  const dispatch = useDispatch();
 
   return (
     <ScrollArea withHint>
@@ -199,9 +203,15 @@ export function SelectUseCase({ setUseCase, setOpenedPedagogyModal }: Props) {
                   />
                 }
                 onClick={() => {
-                  track("Onboarding - Restore");
-                  setUseCase(UseCase.recover);
-                  history.push(`/onboarding/${UseCase.recover}/${ScreenId.pairMyNano}`);
+                  track("Onboarding - Restore With Recover");
+
+                  // An array is used here because we'll have to allow Stax later
+                  if (deviceModelId && [DeviceModelId.nanoX].includes(deviceModelId)) {
+                    setUseCase(UseCase.recover);
+                    history.push(`/onboarding/${UseCase.recover}/${ScreenId.pairMyNano}`);
+                  } else {
+                    dispatch(openModal("MODAL_PROTECT_DISCOVER", null));
+                  }
                 }}
               />
             </FeatureToggle>