+++ apps/ledger-live-desktop/src/renderer/modals/UpdateFirmwareModal/steps/02-step-flash-mcu.tsx
@@ -1,24 +1,22 @@
 import React, { useState, useEffect, useMemo } from "react";
-import { useTranslation } from "react-i18next";
 import { timeout } from "rxjs/operators";
 import { Subscriber, from } from "rxjs";
 import styled from "styled-components";
 import { DeviceModelId } from "@ledgerhq/devices";
 import { DeviceInfo, FirmwareUpdateContext } from "@ledgerhq/types-live";
-import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import { hasFinalFirmware } from "@ledgerhq/live-common/hw/hasFinalFirmware";
 import { isDeviceLocalizationSupported } from "@ledgerhq/live-common/manager/localization";
 import firmwareUpdateMain from "@ledgerhq/live-common/hw/firmwareUpdate-main";
 import { withDevicePolling } from "@ledgerhq/live-common/hw/deviceAccess";
 import getDeviceInfo from "@ledgerhq/live-common/hw/getDeviceInfo";
+import { getEnv } from "@ledgerhq/live-common/env";
 import TrackPage from "~/renderer/analytics/TrackPage";
 import Box from "~/renderer/components/Box";
 import FlashMCU from "~/renderer/components/FlashMCU";
-import Installing from "../Installing";
-import { StepProps } from "..";
-import { getEnv } from "@ledgerhq/live-common/env";
 import { mockedEventEmitter } from "~/renderer/components/debug/DebugMock";
+import Installing from "../Installing";
 import { Body as StepUpdatingBody } from "./02-step-updating";
+import { StepProps } from "..";
 
 const Container = styled(Box).attrs(() => ({
   alignItems: "center",
@@ -26,29 +24,32 @@ const Container = styled(Box).attrs(() => ({
   color: "palette.text.shade100",
 }))``;
 
-const Title = styled(Box).attrs(() => ({
-  ff: "Inter|Regular",
-  fontSize: 5,
-  mb: 3,
-}))``;
-
 type BodyProps = {
   installing?: string;
   progress: number;
   deviceModelId: DeviceModelId;
-  firmware: FirmwareUpdateContext;
+  firmware?: FirmwareUpdateContext;
   initialDelayPhase: boolean;
+
+  current: number;
+  total: number;
 };
 
-const Body = ({ installing, progress, firmware, deviceModelId, initialDelayPhase }: BodyProps) => {
-  return installing || !firmware.shouldFlashMCU || initialDelayPhase ? (
-    <Installing installing={installing} progress={progress} />
+const Body = ({
+  installing,
+  progress,
+  firmware,
+  deviceModelId,
+  initialDelayPhase,
+  current,
+  total,
+}: BodyProps) => {
+  return installing || !firmware?.shouldFlashMCU || initialDelayPhase ? (
+    <Installing installing={installing} current={current} total={total} progress={progress} />
   ) : (
     <FlashMCU deviceModelId={deviceModelId} />
   );
 };
-
-type Props = StepProps;
 
 const DELAY_PHASE = 10000;
 
@@ -58,17 +59,20 @@ const StepFlashMcu = ({
   setError,
   transitionTo,
   setUpdatedDeviceInfo,
-}: Props) => {
-  const { t } = useTranslation();
+}: StepProps) => {
   const [installing, setInstalling] = useState<string | undefined>(undefined);
   const [initialDelayPhase, setInitialDelayPhase] = useState(true);
+
+  // In order to be able to show a progress wording such as 'Installing update 1 of 3'
+  const [maxSteps, setMaxSteps] = useState(0);
+  const [currentStep, setCurrentStep] = useState(0);
+
   // when autoUpdatingMode is true, we simply display the same content as in "step-updating" as the device turns into auto update mode
   const [autoUpdatingMode, setAutoUpdatingMode] = useState(false);
   const [progress, setProgress] = useState(0);
   const withFinal = useMemo(() => hasFinalFirmware(firmware?.final), [firmware]);
 
   const [isMcuUpdateFinished, setIsMcuUpdateFinished] = useState<boolean>(false);
-  const deviceLocalizationFeatureFlag = useFeature("deviceLocalization");
 
   // Gets the updated device info from the command waitForDeviceInfo
   // after a successful MCU update
@@ -88,10 +92,8 @@ const StepFlashMcu = ({
           next: setUpdatedDeviceInfo,
           complete: () => {
             const shouldGoToLanguageStep =
-              firmware &&
-              isDeviceLocalizationSupported(firmware.final.name, deviceModelId) &&
-              deviceLocalizationFeatureFlag?.enabled;
-            transitionTo(shouldGoToLanguageStep ? "deviceLanguage" : "finish");
+              firmware && isDeviceLocalizationSupported(firmware.final.name, deviceModelId);
+            transitionTo(shouldGoToLanguageStep ? "restore" : "finish");
           },
           error: (error: Error) => {
             setError(error);
@@ -105,18 +107,26 @@ const StepFlashMcu = ({
         sub.unsubscribe();
       }
     };
-  }, [
-    deviceLocalizationFeatureFlag?.enabled,
-    deviceModelId,
-    firmware,
-    isMcuUpdateFinished,
-    setError,
-    setUpdatedDeviceInfo,
-    transitionTo,
-  ]);
+  }, [deviceModelId, firmware, isMcuUpdateFinished, setError, setUpdatedDeviceInfo, transitionTo]);
+
+  useEffect(() => {
+    if (!installing) return;
+
+    // Whenever the 'installing' state changes we are flashing a different step,
+    // leverage this to guess the number of steps and which one we are on.
+    setMaxSteps(maxSteps => {
+      if (maxSteps) return maxSteps; // If we've already set a max step, we rely on that.
+      return installing === "flash-bootloader" ? 3 : installing === "flash-mcu" ? 2 : 1;
+    });
+
+    // A change here also means we've moved to a new step, so increase the counter (?)
+    setCurrentStep(currentStep => currentStep + 1);
+  }, [installing]);
 
   // Updates the MCU
   useEffect(() => {
+    if (!firmware) return;
+
     setTimeout(() => {
       setInitialDelayPhase(false);
     }, DELAY_PHASE);
@@ -170,12 +180,13 @@ const StepFlashMcu = ({
 
   return (
     <Container data-test-id="firmware-update-flash-mcu-progress">
-      <Title>{installing ? "" : t("manager.modal.mcuTitle")}</Title>
       <TrackPage category="Manager" name="FlashMCU" />
       <Body
         deviceModelId={deviceModelId}
         firmware={firmware}
         installing={installing}
+        current={currentStep}
+        total={maxSteps}
         progress={progress}
         initialDelayPhase={initialDelayPhase}
       />