+++ libs/ledger-live-common/src/families/algorand/bridge/mock.ts
@@ -1,3 +1,5 @@
+// TODO: update path by moving mockHelpers to coin-framework
+
 import { BigNumber } from "bignumber.js";
 import {
   NotEnoughBalance,
@@ -5,23 +7,29 @@ import {
   InvalidAddress,
   FeeTooHigh,
 } from "@ledgerhq/errors";
-import type { AlgorandTransaction, Transaction } from "../types";
+import type {
+  AlgorandTransaction,
+  Transaction,
+} from "@ledgerhq/coin-algorand/types";
 import type {
   Account,
   AccountBridge,
   CurrencyBridge,
 } from "@ledgerhq/types-live";
 import {
+  makeAccountBridgeReceive,
   scanAccounts,
   signOperation,
   broadcast,
   sync,
   isInvalidRecipient,
 } from "../../../bridge/mockHelpers";
-import { getMainAccount } from "../../../account";
-import { makeAccountBridgeReceive } from "../../../bridge/mockHelpers";
-import { assignToAccountRaw, assignFromAccountRaw } from "../serialization";
-import { initAccount } from "../initAccount";
+import { getMainAccount } from "@ledgerhq/coin-framework/account/index";
+import {
+  assignToAccountRaw,
+  assignFromAccountRaw,
+} from "@ledgerhq/coin-algorand/serialization";
+import { initAccount } from "@ledgerhq/coin-algorand/initAccount";
 
 const receive = makeAccountBridgeReceive();